edit

+++ main/prog/3.h
@@ -3,4 +3,4 @@
 //Here it is necessary to enter the order of operation of the program, the numbers from 0 to 63.
 //Example: const uint8_t _3[64] PROGMEM = {6,12,35,0, 18, 22, 53,..n,..44};
 
-const uint8_t _3[64] PROGMEM = {};
+const uint16_t _3[64] PROGMEM = {};
